subdata table done
